Change type of work_done to float

+++ python-libraries/nanover-omni/src/nanover/omni/openmm.py
@@ -233,7 +233,7 @@ class OpenMMSimulationWorkDone(OpenMMSimulation):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.work_done: int = 0
+        self.work_done: float = 0.0
         self.prev_imd_forces: Optional[np.ndarray] = None
         self.prev_imd_indices: Optional[np.ndarray] = None
